bot/requirements.txt
discord.py[voice]
PyNaCl
python-dotenv
opuslib
